orphan_records_handling/requirements.txt
psycopg2-binary
pandas
sqlalchemy